+++ 11BinarySearch/2Optimized.cpp
@@ -5,16 +5,25 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int binarySearchOptimized(const vector<int>& arr, int target) {
-    int l = 0, r = (int)arr.size() - 1;
-    while (l <= r) {
+int binarySearchOptimized(const vector<int> &arr, int target)
+{
+    int l = 0, r = arr.size() - 1;
+    while (l <= r)
+    {
         int m = l + (r - l) / 2;
-        if (arr[m] == target) return m;
-        if (arr[m] < target) l = m + 1;
-        else r = m - 1;
+        if (arr[m] == target)
+            return m;
+        if (arr[m] < target)
+            l = m + 1;
+        else
+            r = m - 1;
     }
     return -1;
 }
 
-int main() { vector<int> arr = {1, 3, 5, 7, 9}; cout << binarySearchOptimized(arr, 7) << "\n"; return 0; }
-
+int main()
+{
+    vector<int> arr = {1, 3, 5, 7, 9};
+    cout << binarySearchOptimized(arr, 7) << "\n";
+    return 0;
+}
