+++ 11BinarySearch/3MoreOptimized.cpp
@@ -5,15 +5,23 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int binarySearchMoreOptimized(const vector<int>& arr, int target) {
-    int l = 0, r = (int)arr.size();
-    while (l < r) {
+int binarySearchMoreOptimized(const vector<int> &arr, int target)
+{
+    int l = 0, r = arr.size();
+    while (l < r)
+    {
         int m = l + (r - l) / 2;
-        if (arr[m] < target) l = m + 1;
-        else r = m;
+        if (arr[m] < target)
+            l = m + 1;
+        else
+            r = m;
     }
     return (l < (int)arr.size() && arr[l] == target) ? l : -1;
 }
 
-int main() { vector<int> arr = {1, 3, 5, 7, 9}; cout << binarySearchMoreOptimized(arr, 7) << "\n"; return 0; }
-
+int main()
+{
+    vector<int> arr = {1, 3, 5, 7, 9};
+    cout << binarySearchMoreOptimized(arr, 7) << "\n";
+    return 0;
+}
