+++ 4PairSum/1BruteForce.cpp
@@ -9,9 +9,9 @@ vector<pair<int, int>> pairSumBrute(const vector<int> &arr, int target)
 {
     vector<pair<int, int>> ans;
 
-    for (int i = 0; i < (int)arr.size(); ++i)
+    for (int i = 0; i < arr.size(); ++i)
     {
-        for (int j = i + 1; j < (int)arr.size(); ++j)
+        for (int j = i + 1; j < arr.size(); ++j)
         {
             if (arr[i] + arr[j] == target)
             {
@@ -26,6 +26,7 @@ vector<pair<int, int>> pairSumBrute(const vector<int> &arr, int target)
 int main()
 {
     vector<int> arr = {1, 4, 45, 6, 10, 8};
+    
     vector<pair<int, int>> ans = pairSumBrute(arr, 14);
 
     for (size_t i = 0; i < ans.size(); ++i)
