+++ 6StockBUY&SELL/3MoreOptimized.cpp
@@ -5,13 +5,20 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int maxProfitMoreOptimized(const vector<int>& prices) {
+int maxProfitMoreOptimized(const vector<int> &prices)
+{
     int best = 0;
-    for (int i = 1; i < (int)prices.size(); ++i) {
-        if (prices[i] > prices[i - 1]) best += prices[i] - prices[i - 1];
+    for (int i = 1; i < prices.size(); ++i)
+    {
+        if (prices[i] > prices[i - 1])
+            best += prices[i] - prices[i - 1];
     }
     return best;
 }
 
-int main() { vector<int> prices = {7, 1, 5, 3, 6, 4}; cout << maxProfitMoreOptimized(prices) << "\n"; return 0; }
-
+int main()
+{
+    vector<int> prices = {7, 1, 5, 3, 6, 4};
+    cout << maxProfitMoreOptimized(prices) << "\n";
+    return 0;
+}
